approle: Add AppRole authentication method
This commit add the ability to authenticate on the approle endpoint.
On a shared concourse ci, you may not want to use a common certificate which will give the same rights to every pipelines/teams.

Using a specific policy for a role allow to have more granular ACL across your pipelines.

To not break default behavior and existing pipelines, I kept the aws-ec2 authentication as default.