token: Expose optionally the vault token to use it on other resources
I needed to get the token to inject it in a terraform variable to configure the vault provider.
Terraform will create a new token (with a ttl) using the token given in parameters.
Unfortunately, I wasn't able to get the token of the vault resource. This patch add a variable to explicitely expose the token in the same path than secret, to be able to read it on other resources